requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.3
Django==3.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.1
inflection==0.5.1
iniconfig==2.0.0
jwt==1.3.1
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
pycparser==2.21
pytest==7.4.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
